Increase timing precision by tracking the fractional part of the start time.
This solves an issue with certain audio formats with very small samples,
that an rounding error would eventually cause the time to drift, and
cause A/V sync issues after long playback.

Now the duration of media samples is calculated as a floating point,
and the fractional part is tracked seperately to not lose any precision.