Skip invalid data that the parser successfully extraced, but the decoder cannot decode.
Previously, the decode routine would fail, and instruct the buffer processor to completly drop the sample. This could however cause issues when our samples are not cut on frame boundaries, as the start of the next frame would potentially be dropped as well, causing the decoding of subsequent frames to fail as well.

In particular, this fixes plaback of MPEG-PS when using the MS MPEG Demultiplexing filter.