Implement SSE2 RGB conversion
This implements chroma upsampling using bilinear interpolation (75:25
vertical, 50:50 horizontal), and RGB conversion of the resulting 4:4:4.

RGB conversion supports the BT.601, BT.709 and SMTPE240M matrices.
The input color range is extracted from the bitstream if available, but defaults
to limited range.
The output color range equals the input range, but support for changing the
range is included and just needs to be exposed to the user.