Improve language tag mapping.
Language tags will now always be mapped onto the "common" ISO 639-2 format.
This works both for the stream languages, as well as the tags configured in LAV Splitter.

This way files which are tagged with the alternate ISO 639-2 tag will be handled more gracefully ("deu"/"ger", "fra"/"fre", etc), as well as files which have a full language name as their language tag.

Fixes issue 79.