Disable the COOK parser
The parser causes problems when decoding COOK from the MPC-HC RM Splitter.
This is technically caused by a short-coming in the processing code, which assumes that the decoder will consume the same amount of data the parser offered to us, and does not re-try consuming it again.

Since the COOK parser does nothing but try to guess the duration, its useless anyway, so disabling it is the easiest solution for now.

Fixes issue #229.