yadif: Add support for changing timebase
In recent versions of the yadif filter, it'll modify the timebase instead of rescaling the timestamps internally.
The avfilter API allows filters to change the timebase to any arbitrary value, so we should not assume that it'll stay the same. Instead, properly rescale the timestamps after the filtering operation back to the DirectShow timebase.

This fixes playback with yadif active after the recent ffmpeg update.