Adjust queue sizes to consume less ram
The queues are only really required to compensate for badly interleaved files, so that subtitles are found early enough.
Initially the problem mostly occured because audio frames had different durations then video frames, so that one queue would starve the other, but by allowing audio queues to grow quite a bit more then video queues, this problem can be avoided.

Instead, we can now limit the queues again, and save a bit of memory in the process.