Avoid needless copying of the encoded audio buffer.
This can lead to drastic performance increase in some formats (most notably MPEG-4 ALS), because the source format will have very large packets that the decoder only consumes very slowly, which means the whole packet would get copied over and over again.

Now, we try to avoid this copy completely if we trust the source to properly pad the data buffer (only LAV for now), or if we don't, we only copy it to the buffer once, and no longer multiple times.