Add support for "Normal", "Impaired" and "NOT" subtitle selector tokens.
Support for more advanced subtitle selectors has been added which allows
selecting hearing impaired streams, as well as explicitly selecting streams
without special flags ("Normal" streams).

The Hearing Impaired selector "h" works exactly like the existing "d" and
"f" selectors, and will only select streams which are properly flagged as
hearing impaired.

The "n" selector for Normal Streams basically has the meaning of "all
streams which are not default, forced or impaired", allowing to build
combinations like "*|nfd" which will select any streams that are either
normal, forced or default, but not hearing impaired.

Additionally, to complement the logic, a negation operator has been added.
The negation operator "!" has to be placed at the beginning of the flags,
and when its present it negates the meaning of the flags.
A combination like "*|!dfh" then means "not default, forced or hearing
impaired", which in turn translates to "*|n", which has the same semantical
meaning.

Fixes issue 233.