Introduce a fallback override mixer mode.
The fallback override mixer will only engage in case the channel layout changes mid-stream and the new layout is not being accepted by the audio renderer. This is typically the case with WMP/WMC, because they do not like channel re-configurations mid-stream.

When the override mixer is activated, it'll mix the audio to the previous channel layout to ensure playback continues.

Fixes issue 50.