msdk_mvc: disabled hardware decoding, it can result in a decoder deadlock
The hardware can get into a state where it constantly reports a "busy"
state, with no chance of gracefully recovering again. This results in
the player basically deadlocking, and needing to be forcefully killed.

Unfortunately this appears to be a recurring issue with Intels decode
API, so until this can be fully confirmed to be resolved, disable
hardware acceleration again.